requirements.txt
anthropic==0.39.0
streamlit==1.39.0
python-docx==1.1.0
python-dotenv==1.0.0
reportlab==4.0.7
